Added tests for checking similarity of destinations, tested at repository level not controller
In this commit I implemented a new way of checking values are given as expected and are changed (or not changed) as expected. This is done by both being able to inject an instance of each repository we need into the test class, and also running raw sql on the in memory database we are using.

#test #story[191]